# BENCH: Improve perf of rolling.Apply.time_rolling