@WillAyd 2019-10-28
This is great thanks for this. Just out of curiosity, what were you aiming for time-wise on these?